Proof of concept for Java website integration. So far this program must be hosted on a server (for testing purposes a local one). You can view the main map screen, a user profile screen as well as the signup and login screens (These need some work to look proper). You can 'sign in' and 'login'. But only shallowly. It just brings you to the profile screen and brings the username along with you. This was simply just to test whether information could be brought from one place to another. More testing is needed and there are still things to be learned. I also had to redesign the structure of the program. Mostly everything is still in the same place relative to each other but there beginning hierarchy had to change to support the server software.
This program was created to test if it would even be possible to run Java and html together. It is possible but it requires a server to be run, at least the way it's currently set up it does. Not sure how well the view part follows clean architecture, but I believe it can be done.